Make Pardot helper functions private and move them to a helper module
